-Added an example of Single Responsibility Principle
-Added an example of Interface segregation principle, where you would split one big interface, to multiple small ones with a clearly defined purpose and naming

-Added an example of Open-Closed Principle.
This was achived by implementing an abstract class Employee, containing an abstract method (with no implementation). Then, the abstract class serves as a base class for creating a Permanent and Temporary Employee classes which derrive from Employee.

In such way, the Employee class is open for extension, but not for modification